+++ stroom-core-server/src/main/java/stroom/connectors/ExternalLibService.java
@@ -34,12 +34,17 @@ public class ExternalLibService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ExternalLibService.class);
 
-    private static final String CONNECTORS_LIB_DIR = "stroom.plugins.lib.dir";
+    private static final String CONNECTORS_LIB_DIR_PROP_KEY = "stroom.plugins.lib.dir";
 
     // A separate class loader will be created for each JAR file found.
     private final Collection<ClassLoader> classLoaders;
 
-    public ExternalLibService(final String connectorsLibDirName) {
+    @Inject
+    public ExternalLibService(final StroomPropertyService propertyService) {
+        this(propertyService.getProperty(CONNECTORS_LIB_DIR_PROP_KEY));
+    }
+
+    private ExternalLibService(final String connectorsLibDirName) {
         // Default to using the current class loader for loading services.
         final List<ClassLoader> classLoaders = new ArrayList<>();
 
@@ -47,28 +52,31 @@ public class ExternalLibService {
             final Path connectorsLibDir = Paths.get(connectorsLibDirName);
             if (Files.isDirectory(connectorsLibDir)) {
                 try (final DirectoryStream<Path> stream = Files.newDirectoryStream(connectorsLibDir, "*.jar")) {
-                    stream.forEach(file -> getClassLoader(file).ifPresent(classLoaders::add));
-                    LOGGER.info("Loaded the External Service JARs from '" + connectorsLibDirName + "'");
+                    stream.forEach(file -> {
+                        getClassLoader(file).ifPresent(classLoader -> {
+                            LOGGER.info("Adding external jar {}", file.getFileName().toString());
+                            classLoaders.add(classLoader);
+                        });
+                    });
+                    LOGGER.info("Loaded the External Service JARs from '{}'", connectorsLibDirName);
                 } catch (final IOException e) {
-                    LOGGER.warn("Could not load External Service JARs from '" + connectorsLibDirName + "'", e);
+                    LOGGER.error("Could not load External Service JARs from '{}'", connectorsLibDirName, e);
                 }
             } else {
-                LOGGER.warn("Could not load External Service JARs from '" + connectorsLibDirName + "' as dir does not exist");
+                LOGGER.warn("Could not load External Service JARs from '{}' as dir does not exist", connectorsLibDirName);
             }
+        } else {
+            LOGGER.warn("Property {} is not set, no external Jars can be loaded", CONNECTORS_LIB_DIR_PROP_KEY);
         }
 
         this.classLoaders = classLoaders;
     }
 
-    @Inject
-    public ExternalLibService(final StroomPropertyService propertyService) {
-        this(propertyService.getProperty(CONNECTORS_LIB_DIR));
-    }
 
     /**
      * Wraps the call to URL, catches and prints the exception.
      *
-     * @param uri The URI to call toURL() on.
+     * @param path The path to convert to a {@link URLClassLoader}
      * @return The URL from the URI, or null if that fails.
      */
     private Optional<URLClassLoader> getClassLoader(final Path path) {
@@ -78,10 +86,7 @@ public class ExternalLibService {
             urlClassLoader = new URLClassLoader(new URL[]{url});
 
         } catch (final MalformedURLException e) {
-            final String msg = String.format("Could not load in an External Service JAR %s: %s",
-                    path.toString(),
-                    e.getLocalizedMessage());
-            LOGGER.error(msg, e);
+            LOGGER.error("Could not load in an External Service JAR {}", path.toString(), e);
         } catch (final Exception e) {
             LOGGER.error(e.getMessage(), e);
         }