+++ stroom-kafka-client-impl_0_10_0_1/src/main/java/stroom/connectors/kafka/StroomKafkaProducerImpl.java
@@ -5,16 +5,17 @@ import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import stroom.connectors.ConnectorProperties;
 
-import java.io.IOException;
-import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Properties;
-import java.util.concurrent.Future;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 /**
  * A singleton responsible for sending records to Kafka.
@@ -28,12 +29,10 @@ class StroomKafkaProducerImpl implements StroomKafkaProducer {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(StroomKafkaProducerImpl.class);
 
-    private static final int TIME_BETWEEN_INIT_ATTEMPS_MS = 30_000;
     private final ConnectorProperties properties;
     private final String bootstrapServers;
     //instance of a kafka producer that will be shared by all threads
-    private volatile Producer<String, byte[]> producer = null;
-    private volatile Instant timeOfLastFailedInitAttempt = Instant.EPOCH;
+    private final Producer<String, byte[]> producer;
 
     public StroomKafkaProducerImpl(final ConnectorProperties properties) {
         this.properties = properties;
@@ -46,104 +45,113 @@ class StroomKafkaProducerImpl implements StroomKafkaProducer {
                 final String msg = String.format(
                         "Stroom is not properly configured to connect to Kafka: %s is required.",
                         StroomKafkaProducer.BOOTSTRAP_SERVERS_CONFIG);
-                LOGGER.error(msg);
+                throw new RuntimeException(msg);
             }
         } else {
             final String msg = String.format(
                     "Stroom is not properly configured to connect to Kafka: properties containing at least %s are required.",
                     StroomKafkaProducer.BOOTSTRAP_SERVERS_CONFIG);
-            LOGGER.error(msg);
+            throw new RuntimeException(msg);
         }
+
+        LOGGER.info("Initialising kafka producer for {}", bootstrapServers);
+        Properties props = getProperties();
+
+        LOGGER.info("Creating Kafka Producer with Props: " + props);
+
+        final ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
+        try {
+            // This bizzarness has to be done to force the Kafka config libraries to use the 'Kafka Classloader'
+            // instead of the current thread content class loader.
+            // https://issues.apache.org/jira/browse/KAFKA-3218
+            Thread.currentThread().setContextClassLoader(null);
+
+            //TODO https://community.hortonworks.com/articles/80813/kafka-best-practices-1.html suggests that
+            //having 1 producer per topic yields better performance. We could keep a map of producers keyed
+            //on topic, though we would need to kill off producers that were unused to cope with users mistyping
+            //topic names.  This could probably only be done once StroomKafkaProducer is separated from the
+            //Connector interface
+
+            //Even if kafka is down the producer will still create successfully. Any calls to send will however
+            //block until it comes up or throw an exception on timeout
+            this.producer = new KafkaProducer<>(props);
+        } catch (Exception e) {
+            throw new RuntimeException(String.format("Error initialising kafka producer for %s, due to %s",
+                    bootstrapServers, e.getMessage()), e);
+        } finally {
+            Thread.currentThread().setContextClassLoader(classLoader);
+        }
+        LOGGER.info("Kafka Producer successfully created (this doesn't guarantee Kafka is available)");
     }
 
-    public void send(final StroomKafkaProducerRecord<String, byte[]> stroomRecord,
-                     final boolean flushOnSend,
-                     final Consumer<Exception> exceptionHandler) {
 
-        //kafka may not have been up on startup so ensure we have a producer instance now
-        try {
-            initProducer();
-        } catch (Exception e) {
-            final String msg = String.format("Error initialising kafka producer to %s, (%s)",
-                    this.bootstrapServers,
-                    e.getMessage());
-            LOGGER.error(msg);
-            exceptionHandler.accept(e);
-            return;
-        }
+    @Override
+    public List<CompletableFuture<StroomKafkaRecordMetaData>> sendAsync(
+            final List<StroomKafkaProducerRecord<String, byte[]>> stroomRecords,
+            final Consumer<Throwable> exceptionHandler) {
 
-        if (producer != null) {
+        return send(stroomRecords, exceptionHandler);
+    }
+
+    @Override
+    public List<StroomKafkaRecordMetaData> sendSync(final List<StroomKafkaProducerRecord<String, byte[]>> stroomRecords) {
+
+        List<CompletableFuture<StroomKafkaRecordMetaData>> futures = send(stroomRecords, null);
+
+        List<StroomKafkaRecordMetaData> metaDataList = new ArrayList<>();
+        //Now wait for them all to complete
+        futures.forEach(future -> {
             try {
-                final ProducerRecord<String, byte[]> record =
-                        new ProducerRecord<>(
-                                stroomRecord.topic(),
-                                stroomRecord.partition(),
-                                stroomRecord.timestamp(),
-                                stroomRecord.key(),
-                                stroomRecord.value());
-
-                Future<RecordMetadata> future = producer.send(record, (recordMetadata, exception) -> {
-                    if (exception != null) {
-                        exceptionHandler.accept(exception);
-                    }
-                    LOGGER.trace("Record sent to Kafka");
-                });
-
-                if (flushOnSend) {
-                    future.get();
-                }
-
-            } catch (Exception e) {
-                final String msg = String.format("Error initialising kafka producer to %s, (%s)",
-                        this.bootstrapServers,
-                        e.getMessage());
-                LOGGER.error(msg);
-                exceptionHandler.accept(e);
+                metaDataList.add(future.get());
+            } catch (InterruptedException e) {
+                LOGGER.warn("Thread {} interrupted", Thread.currentThread().getName());
+                Thread.currentThread().interrupt();
+            } catch (ExecutionException e) {
+                //this is sync so throw rather than using the callback
+                throw new RuntimeException(
+                        String.format("Error sending %s records to kafka", futures.size()), e);
             }
-        } else {
-            final String msg = String.format("Kafka producer is currently not initialised, " +
-                            "it may be down or the connection details incorrect, bootstrapServers: [%s]",
-                    this.bootstrapServers);
-            exceptionHandler.accept(new IOException(msg));
-        }
+        });
+        return metaDataList;
+    }
+
+    private List<CompletableFuture<StroomKafkaRecordMetaData>> send(
+            final List<StroomKafkaProducerRecord<String, byte[]>> stroomRecords,
+            final Consumer<Throwable> exceptionHandler) {
+
+        List<CompletableFuture<StroomKafkaRecordMetaData>> futures = stroomRecords.stream()
+                .map(this::mapStroomRecord)
+                .map(producerRecord -> {
+                    CompletableFuture<StroomKafkaRecordMetaData> future = new CompletableFuture<>();
+                    producer.send(producerRecord, (recordMetadata, exception) -> {
+                        if (exception != null) {
+                            future.completeExceptionally(exception);
+                            if (exceptionHandler != null) {
+                                exceptionHandler.accept(exception);
+                            }
+                        } else {
+                            future.complete(WrappedRecordMetaData.wrap(recordMetadata));
+                        }
+                        LOGGER.trace("Record sent to Kafka");
+                    });
+                    return future;
+                })
+                .collect(Collectors.toList());
+
+        return futures;
+    }
+
+    private ProducerRecord<String, byte[]> mapStroomRecord(final StroomKafkaProducerRecord<String, byte[]> stroomRecord) {
+        return new ProducerRecord<>(
+                stroomRecord.topic(),
+                stroomRecord.partition(),
+                stroomRecord.timestamp(),
+                stroomRecord.key(),
+                stroomRecord.value());
     }
 
     public void flush() {
-        if (producer != null) {
-            producer.flush();
-        }
-    }
-
-    /**
-     * Lazy initialisation of the kafka producer
-     */
-    private void initProducer() {
-        if (producer == null && isOkToInitNow(bootstrapServers)) {
-            synchronized (this) {
-                if (producer == null && isOkToInitNow(bootstrapServers)) {
-                    LOGGER.info("Initialising kafka producer for {}", bootstrapServers);
-                    Properties props = getProperties();
-
-                    LOGGER.info("Creating Kafka Producer with Props: " + props);
-
-                    final ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-                    try {
-                        // This bizzarness has to be done to force the Kafka config libraries to use the 'Kafka Classloader'
-                        // instead of the current thread content class loader.
-                        // https://issues.apache.org/jira/browse/KAFKA-3218
-                        Thread.currentThread().setContextClassLoader(null);
-                        producer = new KafkaProducer<>(props);
-                    } catch (Exception e) {
-                        LOGGER.error("Error initialising kafka producer for {}", bootstrapServers, e);
-                        throw e;
-                    } finally {
-                        Thread.currentThread().setContextClassLoader(classLoader);
-                    }
-                    LOGGER.info("Kafka Producer successfully created");
-                    LOGGER.info(String.format("With Properties: %s", props));
-                }
-            }
-        }
+        producer.flush();
     }
 
     private Properties getProperties() {
@@ -166,25 +174,13 @@ class StroomKafkaProducerImpl implements StroomKafkaProducer {
         return props;
     }
 
-    /**
-     * A check to prevent many process trying to repeatedly init a producer. This means init will only be attempted
-     * every 30s to prevent the logs being filled up if kafka is down. It will never try to init if there is no
-     * bootstrapServers value
-     */
-    private boolean isOkToInitNow(final String bootstrapServers) {
-        return !Strings.isNullOrEmpty(bootstrapServers) &&
-                Instant.now()
-                        .isAfter(timeOfLastFailedInitAttempt.plusMillis(TIME_BETWEEN_INIT_ATTEMPS_MS));
-    }
-
+    @Override
     public void shutdown() {
-        if (producer != null) {
-            try {
-                LOGGER.info("Closing down Kafka Producer");
-                producer.close();
-            } catch (Exception e) {
-                LOGGER.error("Error closing kafka producer", e);
-            }
+        try {
+            LOGGER.info("Closing down Kafka Producer");
+            producer.close();
+        } catch (Exception e) {
+            LOGGER.error("Error closing kafka producer", e);
         }
     }
 }