+++ stroom-kafka-client-impl_0_10_0_1/src/test/java/stroom/connectors/kafka/TestStroomKafkaProducer.java
@@ -9,6 +9,7 @@ import stroom.connectors.ConnectorProperties;
 import stroom.connectors.ConnectorPropertiesEmptyImpl;
 
 import java.nio.charset.StandardCharsets;
+import java.util.Collections;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Consumer;
 
@@ -27,7 +28,7 @@ public class TestStroomKafkaProducer {
         StroomKafkaProducerFactoryImpl stroomKafkaProducerFactory = new StroomKafkaProducerFactoryImpl();
         ConnectorProperties kafkaProps = new ConnectorPropertiesEmptyImpl();
         kafkaProps.put(StroomKafkaProducer.BOOTSTRAP_SERVERS_CONFIG, "kafka:9092");
-        StroomKafkaProducer stroomKafkaProducer = stroomKafkaProducerFactory.getConnector(KAFKA_VERSION, kafkaProps);
+        StroomKafkaProducer stroomKafkaProducer = stroomKafkaProducerFactory.createConnector(KAFKA_VERSION, kafkaProps);
         StroomKafkaProducerRecord<String, byte[]> record =
                 new StroomKafkaProducerRecord.Builder<String, byte[]>()
                         .topic("statistics")
@@ -36,17 +37,17 @@ public class TestStroomKafkaProducer {
                         .build();
 
         // When
-        stroomKafkaProducer.send(record, true, DEFAULT_CALLBACK);
+        stroomKafkaProducer.sendSync(Collections.singletonList(record));
 
         // Then: manually check your Kafka instances 'statistics' topic for 'some record data'
     }
 
-    @Test
+    @Test(expected = RuntimeException.class)
     public void testBadlyConfigured() {
         // Given
         StroomKafkaProducerFactoryImpl stroomKafkaProducerFactory = new StroomKafkaProducerFactoryImpl();
         ConnectorProperties properties = new ConnectorPropertiesEmptyImpl();
-        StroomKafkaProducer stroomKafkaProducer = stroomKafkaProducerFactory.getConnector(KAFKA_VERSION, properties);
+        StroomKafkaProducer stroomKafkaProducer = stroomKafkaProducerFactory.createConnector(KAFKA_VERSION, properties);
         StroomKafkaProducerRecord<String, byte[]> record =
                 new StroomKafkaProducerRecord.Builder<String, byte[]>()
                         .topic("statistics")
@@ -57,7 +58,7 @@ public class TestStroomKafkaProducer {
         AtomicBoolean hasSendFailed = new AtomicBoolean(false);
 
         // When
-        stroomKafkaProducer.send(record, true, ex -> hasSendFailed.set(true));
+        stroomKafkaProducer.sendSync(Collections.singletonList(record));
 
         // Then
         Assert.assertTrue(hasSendFailed.get());