+++ stroom-statistics-server/src/main/java/stroom/statistics/server/stroomstats/internal/StroomStatsInternalStatisticsService.java
@@ -32,6 +32,7 @@ import java.util.GregorianCalendar;
 import java.util.List;
 import java.util.Map;
 import java.util.TimeZone;
+import java.util.function.Consumer;
 
 @SuppressWarnings("unused")
 @Component
@@ -44,17 +45,17 @@ class StroomStatsInternalStatisticsService implements InternalStatisticsService 
     private static final Class<Statistics> STATISTICS_CLASS = Statistics.class;
     private static final TimeZone TIME_ZONE_UTC = TimeZone.getTimeZone(ZoneId.from(ZoneOffset.UTC));
 
-    private final StroomKafkaProducer stroomKafkaProducer;
+    private final StroomKafkaProducerFactoryService stroomKafkaProducerFactory;
     private final StroomPropertyService stroomPropertyService;
     private final String docRefType;
     private final JAXBContext jaxbContext;
     private final DatatypeFactory datatypeFactory;
 
-    // If we move to a 'named' kafka config later, change the java.util.function.Supplier to a java.util.function.Function
-    StroomStatsInternalStatisticsService(final StroomKafkaProducer stroomKafkaProducer,
+    @Inject
+    StroomStatsInternalStatisticsService(final StroomKafkaProducerFactoryService stroomKafkaProducerFactory,
                                          final StroomPropertyService stroomPropertyService) {
+        this.stroomKafkaProducerFactory = stroomKafkaProducerFactory;
         this.stroomPropertyService = stroomPropertyService;
-        this.stroomKafkaProducer = stroomKafkaProducer;
         this.docRefType = stroomPropertyService.getProperty(PROP_KEY_DOC_REF_TYPE);
 
         try {
@@ -71,47 +72,41 @@ class StroomStatsInternalStatisticsService implements InternalStatisticsService 
         }
     }
 
-    @Inject
-    StroomStatsInternalStatisticsService(final StroomKafkaProducerFactoryService stroomKafkaProducerFactory,
-                                         final StroomPropertyService stroomPropertyService) {
-        this(stroomKafkaProducerFactory.getProducer(exception -> LOGGER.error("Unable to call function on Kafka producer!", exception))
-                , stroomPropertyService);
-    }
-
     @Override
     public void putEvents(final Map<DocRef, List<InternalStatisticEvent>> eventsMap) {
 
+        StroomKafkaProducer stroomKafkaProducer = stroomKafkaProducerFactory.getConnector().orElse(null);
+        if (stroomKafkaProducer == null) {
+            throw new RuntimeException("The Kafka producer isn't initialised, unable to send any events");
+        }
+
         Preconditions.checkNotNull(eventsMap);
 
-        try {
-            //We work on the basis that a stat may or may not have a valid datasource (StatisticConfiguration) but we
-            //will let stroom-stats worry about that and just fire what we have at kafka
-            eventsMap.entrySet().stream()
-                    .filter(entry ->
-                            !entry.getValue().isEmpty())
-                    .forEach(entry -> {
-                        DocRef docRef = entry.getKey();
-                        List<InternalStatisticEvent> events = entry.getValue();
-                        String statName = docRef.getName();
-                        //all have same name so have same type
-                        String topic = getTopic(events.get(0).getType());
-                        byte[] message = buildMessage(docRef, events);
-                        StroomKafkaProducerRecord<String, byte[]> producerRecord =
-                                new StroomKafkaProducerRecord.Builder<String, byte[]>()
-                                        .topic(topic)
-                                        .key(docRef.getUuid())
-                                        .value(message)
-                                        .build();
-                        stroomKafkaProducer.send(producerRecord, false, exception -> {
-                            throw new RuntimeException(String.format(
-                                    "Error sending %s internal stats with name %s to kafka on topic %s, due to (%s)",
-                                    events.size(), statName, topic, exception.getMessage()), exception);
-                        });
-                    });
-        } finally {
-            //any problems with the send will call the exceptionhandler above to be triggered at this point
-            stroomKafkaProducer.flush();
-        }
+        //We work on the basis that a stat may or may not have a valid datasource (StatisticConfiguration) but we
+        //will let stroom-stats worry about that and just fire what we have at kafka
+        eventsMap.entrySet().stream()
+                .filter(entry ->
+                        !entry.getValue().isEmpty())
+                .forEach(entry -> {
+                    DocRef docRef = entry.getKey();
+                    List<InternalStatisticEvent> events = entry.getValue();
+                    String statName = docRef.getName();
+                    //all have same name so have same type
+                    String topic = getTopic(events.get(0).getType());
+                    byte[] message = buildMessage(docRef, events);
+                    String key = docRef.getUuid();
+                    StroomKafkaProducerRecord<String, byte[]> producerRecord =
+                            new StroomKafkaProducerRecord.Builder<String, byte[]>()
+                                    .topic(topic)
+                                    .key(key)
+                                    .value(message)
+                                    .build();
+                    Consumer<Throwable> exceptionHandler = StroomKafkaProducer
+                            .createLogOnlyExceptionHandler(LOGGER, topic, key);
+
+                    //These are only internal stats so just send them async for performance
+                    stroomKafkaProducer.sendAsync(producerRecord, exceptionHandler);
+                });
     }
 
     private byte[] buildMessage(final DocRef docRef, final List<InternalStatisticEvent> events) {